Add files via upload
Apologies, for the website project the following files. views.py, app.py, index.html, profile.html, and index.js are used to create the web page. I wasn't able to upload a folder which led to me manually uploading micro files, for the website. If you are using the website project for inspiration and it does not seem to be working please message me, I will happily help you out. Happy coding!